requirements.txt
pillow
pytz
celery
pyexcel
pyexcel-xls
pyexcel-xlsx
pycrypto
redis